+++ auth/src/auth/repository/auth.repository.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { User, UserDocument } from './user.shema';
-import { UserRole } from './user.role';
+import { User, UserDocument } from '../schema/user.shema';
+import { UserRole } from '../../common/enums/user.role';
 
 @Injectable()
 export class AuthRepository {
