+++ auth/src/auth/service/auth.service.ts
@@ -13,25 +13,25 @@ import { ConfigService } from '@nestjs/config';
 import { randomUUID } from 'crypto';
 import { ClientKafka } from '@nestjs/microservices';
 
-import { AuthRepository } from './auth.repository';
-import { UserRole } from './user.role';
-import { UserDocument } from './user.shema';
+import { AuthRepository } from '../repository/auth.repository';
+import { UserRole } from '../../common/enums/user.role';
+import { UserDocument } from '../schema/user.shema';
 
-import { RegisterRequest } from './dto/request/register.request';
-import { LoginRequest } from './dto/request/login.request';
-import { UpdateInfoRequest } from './dto/request/update.info.request';
-import { AssignRoleRequest } from './dto/request/assign.role.request';
+import { RegisterRequest } from '../dto/request/register.request';
+import { LoginRequest } from '../dto/request/login.request';
+import { UpdateInfoRequest } from '../dto/request/update.info.request';
+import { AssignRoleRequest } from '../dto/request/assign.role.request';
 
-import { DuplicateResponse } from './dto/response/dupblicate.response';
-import { RegisterResponse } from './dto/response/register.response';
-import { LoginResponse } from './dto/response/login.response';
-import { BasicResponse } from './dto/response/basic.response';
-import { GetProfileResponse } from './dto/response/get.profile.response';
-import { AllUserResponse } from './dto/response/all.user.response';
-import { UserDto } from './dto/response/use.dto';
+import { DuplicateResponse } from '../dto/response/dupblicate.response';
+import { RegisterResponse } from '../dto/response/register.response';
+import { LoginResponse } from '../dto/response/login.response';
+import { BasicResponse } from '../../common/responses/basic.response';
+import { GetProfileResponse } from '../dto/response/get.profile.response';
+import { AllUserResponse } from '../dto/response/all.user.response';
+import { UserDto } from '../dto/response/user.dto';
 
 import { Request } from 'express';
-import { JwtUser } from './dto/user.decorator';
+import { JwtUser } from '../../common/decorators/user.decorator';
 
 /**
  * 인증(Authentication) 관련 비즈니스 로직을 처리하는 서비스입니다.
