+++ auth/src/common/responses/basic.response.ts
